+++ src/components/ML/Cameras.tsx
@@ -1,10 +1,23 @@
+import type { VRM } from "@pixiv/three-vrm";
+import { VRMContext } from "context";
+import { useContext } from "react";
 import Webcam from "react-webcam";
 import { state } from "state";
 import { useSnapshot } from "valtio";
-// import { activateDraw } from "./utils";
+import { handleHandtrack } from "./handtrack.setup";
+import { handleHolistic } from "./holistic.setup";
+
+const activateDraw = (ref: HTMLCanvasElement, vrm: { current: VRM }) => {
+  const hollisticInput = document.querySelector(".input_video");
+  const handtrackInput = document.querySelector(".input_video2");
+
+  handleHolistic(vrm, hollisticInput as HTMLVideoElement);
+  handleHandtrack(handtrackInput as HTMLVideoElement);
+};
 
 export default function Cameras() {
   const { cameraStarted } = useSnapshot(state);
+  const { vrm } = useContext(VRMContext);
   return (
     <>
       {cameraStarted ? (
@@ -20,7 +33,10 @@ export default function Cameras() {
               state.cameraLoaded = true;
             }}
           />
-          {/* <canvas className="guides" ref={(e: any) => activateDraw(e)}></canvas> */}
+          <canvas
+            className="guides"
+            ref={(e: any) => activateDraw(e, vrm)}
+          ></canvas>
         </>
       ) : null}
     </>
